Add files via upload
Digital Clock is a C++ application that displays the current time and date in a digital format. It utilizes the built-in functions for date and time manipulation provided by the standard library. The application continuously updates the clock display to show the accurate time, ensuring that users have an easy and convenient way to keep track of the current time and date.